+++ youshan/models/user.py
@@ -44,10 +44,5 @@ class User(Base):
         if hasattr(member, 'display_name'):
             self.display_name = member.display_name
 
-        self.getUUID()
-
     def __repr__(self):
-        return f'<User instance of {self.nick_name}, puid {self.puid}>'
-
-    def getUUID(self):
-        self.uuid = self.r.hget(self.group.uuid, self.nick_name)
+        return f'<User instance of {self.nick_name}>'
